+++ CHANGELOG.md
@@ -3,4 +3,8 @@ All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
 ## 2020 - 2020/02/19
-- first version. Cat image to terminal in colour or greyscale
+- First version. Cat image to terminal in colour or greyscale
+
+## 2020.1 - 2020/02/20
+- Removed unused variables, cleaned up return statement, reformat
+parser.add_argument